alt_ml_pipeline_1/requirements.txt
# Alt ML Pipeline 1 - Requirements
# Python 3.11 recommended for best ML library compatibility
# Install with: pip install -r requirements.txt

# Core Deep Learning (Python 3.11 compatible versions)
torch==2.1.2
torchvision==0.16.2
pytorch-lightning==2.1.3

# Computer Vision & YOLO
ultralytics==8.0.227
opencv-python==4.8.1.78
albumentations==1.3.1
timm==0.9.12

# Image Processing
Pillow==10.0.1
scikit-image==0.22.0
imageio==2.31.6
tifffile==2023.9.26

# Data Science
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
tensorboard==2.15.1

# Experiment Tracking
mlflow==2.9.1
wandb==0.16.0

# Augmentation & Transforms
kornia==0.7.0
imgaug==0.4.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
jupyter==1.0.0
ipywidgets==8.1.1

# AWS Integration
boto3>=1.33.3,<2.0
sagemaker==2.199.0

# Optional: Advanced cell segmentation
# cellpose==2.2.3  # Uncomment if using transfer learning
